slight refactor, writer to allow file handle input
