+++ fastnml/__init__.py
@@ -1,13 +1,18 @@
-__version__ = '0.1.1'
+__version__ = '0.1.2'
 
 import f90nml  # noqa: E402
 assert f90nml
 
-from .process import (
+from .reader import (
     read_namelist,
-    save_namelist,
 )  # noqa: E402
 assert all((
     read_namelist,
+))
+
+from .writer import (
+    save_namelist,
+)  # noqa: E402
+assert all((
     save_namelist,
 ))
